requirements.txt
numpy==1.16.4
pygame==1.9.6
PyOpenGL==3.1.0
